refactor sidenav
