+++ desktop/src/renderer/src/components/SideNav/NavItem.d.tsx
@@ -1,0 +1,8 @@
+import type { ReactNode } from 'react';
+
+export type NavItemProps = {
+	path: string;
+	label: string;
+	icon: ReactNode;
+	role?: string;
+};
