+++ desktop/src/renderer/src/components/SideNav/index.ts
@@ -1,0 +1,1 @@
+export * from './SideNav';
